Improves first-time experience by switching default to MySQL
MySQL is the best environment for people to learn zipkin. It is widely
understood, can be easily explored using SQL, supports on-demand
dependency aggregation, returns precise results for queries, and has
had very few support issues.

While we want to show people how to work with Cassandra at some point,
it has proven a high barrier entry for the opposite of above. People
rarely know cassandra intimately, it has imprecise results (usually
users have to over-fetch), the duration query doesn't work, the
dependency graph requires a cron, and it has had numerous support
issues.

The point of our docker-compose file is to ease the 5 minute, getting
started story of Zipkin. At the moment, MySQL is by far the best choice
for this and is now the default.