Switch to download zipkin-autoconfigure-collector-kafak10-x.x.x-module.jar from public repository, instead of using contents checking into the project.
One sticking point was that the fat jars produced by the spring boot maven plugin with the module layout seemed to have some incompatibility with the version of unzip installed in this image by default. Using unzip to extract or list the contents of one of these jars produces an error message - "short read". The solution (borrowed from https://github.com/openzipkin/docker-zipkin-aws/blob/master/Dockerfile) was to install and remove a different version of unzip using apk, using unzip to expand the jar in between.